Move wrapper classes to core
