+++ spikeinterface/sorters/utils/__init__.py
@@ -1,2 +1,2 @@
 from .shellscript import ShellScript
-from .misc import (SpikeSortingError, get_git_commit, RecordingExtractorOldAPI)
+from .misc import (SpikeSortingError, get_git_commit)
